backend/requirements.txt
fastapi
gunicorn
uvicorn
pydantic
motor
python-dotenv
azure-cosmos
requests
beautifulsoup4
lxml
azure-ai-evaluation
langchain
langchain_openai
langchain_community
azure-identity
pydantic[email]
pydantic-settings
passlib
PyJWT